Move useGetSettings into own import
