+++ gsa/src/web/utils/useGetSettings.js
@@ -1,0 +1,53 @@
+/* Copyright (C) 2020 Greenbone Networks GmbH
+ *
+ * SPDX-License-Identifier: GPL-2.0-or-later
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU General Public License
+ * as published by the Free Software Foundation; either version 2
+ * of the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program; if not, write to the Free Software
+ * Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA.
+ */
+import gql from 'graphql-tag';
+
+import {useQuery} from '@apollo/react-hooks';
+
+import {toFruitfulQuery} from 'web/utils/graphql';
+
+export const GET_SETTING = gql`
+  query UserSetting($userSettingId: UUID!) {
+    userSetting(userSettingId: $userSettingId) {
+      uuid
+      name
+      value
+      comment
+    }
+  }
+`;
+
+export const useGetSetting = () => {
+  return toFruitfulQuery(useQuery)(GET_SETTING);
+};
+
+export const GET_SETTINGS = gql`
+  query UserSettings($filterString: String) {
+    userSettings(filterString: $filterString) {
+      uuid
+      name
+      value
+      comment
+    }
+  }
+`;
+
+export const useGetSettings = () => {
+  return toFruitfulQuery(useQuery)(GET_SETTINGS);
+};
